# packages/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: StreamYolo API
  version: 0.0.1
  description: Room-first live creator platform — rooms, tips, private sessions, tokens, CCBill.
servers:
  - url: http://localhost:3001
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token — delivered via Set-Cookie on login; also accepted as Bearer header for native clients.

  # ─── Reusable parameters ───────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Insufficient role or ownership
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
...
    PublicUser:
      type: object
      required: [id, displayName, role, status, createdAt]
      properties:
        id: { type: string }
        displayName: { type: string }
        username: { type: string }
        role:
          type: string
          enum: [VIEWER, CREATOR, ADMIN, MODERATOR]
        status:
          type: string
          enum: [ACTIVE, SUSPENDED, DELETED]
        createdAt: { type: string, format: date-time }
    WalletSummary:
      type: object
      required: [tokenBalance, reservedTokenBalance, lifetimePurchasedTokens, lifetimeSpentTokens]
      properties:
        tokenBalance: { type: integer }
        reservedTokenBalance: { type: integer }
        lifetimePurchasedTokens: { type: integer }
        lifetimeSpentTokens: { type: integer }
...
    CreatorSummary:
      type: object
      required: [id, userId, displayName, isLive, user]
      properties:
        id: { type: string }
        userId: { type: string }
        displayName: { type: string }
        user: { $ref: '#/components/schemas/PublicUser' }
        avatarUrl:
          type: string
          nullable: true
        isLive: { type: boolean }
        status:
          type: string
          enum: [DRAFT, PENDING, ACTIVE, PAUSED, SUSPENDED, BANNED]
        bio:
          type: string
          nullable: true
        privateRateTokensPerMinute:
          type: integer
          nullable: true
        minPrivateMinutes:
          type: integer
          nullable: true
        privateViewerCamRequired:
          type: boolean
          nullable: true
        privateScreenShareAllowed:
          type: boolean
          nullable: true
        privateRulesText:
          type: string
          nullable: true
...
    RoomTagDto:
      type: object
      required: [slug, label]
      properties:
        slug: { type: string }
        label: { type: string }
        group:
          type: string
          nullable: true
...
    RoomDetail:
      type: object
      required: [id, title, status, visibility, creator, viewerCount, activePrivateSessionId, privateRateTokensPerMinute, minPrivateMinutes, privateViewerCamRequired, privateScreenShareAllowed, tags, createdAt]
      properties:
        id: { type: string }
        title: { type: string }
        status:
          type: string
          enum: [DRAFT, LIVE, ENDED, SUSPENDED, HIDDEN]
        visibility:
          type: string
          enum: [PUBLIC, UNLISTED]
        category:
          type: string
          nullable: true
          description: MALE, FEMALE, COUPLES, or TRANS when set
        countryCode:
          type: string
          minLength: 2
          maxLength: 2
          nullable: true
        countryName:
          type: string
          nullable: true
        tags:
          type: array
          items: { $ref: '#/components/schemas/RoomTagDto' }
        creator: { $ref: '#/components/schemas/CreatorSummary' }
        thumbnailUrl:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        viewerCount: { type: integer }
        activePrivateSessionId:
          type: string
          nullable: true
        privateRateTokensPerMinute: { type: integer }
        minPrivateMinutes: { type: integer }
        privateViewerCamRequired: { type: boolean }
        privateScreenShareAllowed: { type: boolean }
        privateRulesText:
          type: string
          nullable: true
        goal:
          type: object
          allOf:
            - { $ref: '#/components/schemas/RoomGoalDto' }
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        endedAt:
          type: string
          format: date-time
          nullable: true
        createdAt: { type: string, format: date-time }
...
    RoomGoalDto:
      type: object
      required: [id, title, targetTokens, currentTokens]
      properties:
        id: { type: string }
        title: { type: string }
        targetTokens: { type: integer }
        currentTokens: { type: integer }

    # ── Creator menu items ────────────────────────────────────────────────────
...
    PrivateSessionDto:
      type: object
      required: [id, creatorId, viewerId, publicRoomId, status, rateTokensPerMinute, minMinutes, viewerCamRequired, screenShareAllowed, reservedTokens, capturedTokens, releasedTokens, requestedAt, createdAt]
      properties:
        id: { type: string }
        creatorId: { type: string }
        viewerId: { type: string }
        publicRoomId: { type: string }
        status:
          type: string
          enum: [REQUESTED, ACCEPTED, DECLINED, ACTIVE, ENDED, CANCELLED, EXPIRED, FORCE_ENDED]
        rateTokensPerMinute: { type: integer }
        minMinutes: { type: integer }
        viewerCamRequired: { type: boolean }
        screenShareAllowed: { type: boolean }
        rulesText:
          type: string
          nullable: true
        reservedTokens: { type: integer }
        capturedTokens: { type: integer }
        releasedTokens: { type: integer }
        livekitRoomName:
          type: string
          nullable: true
        requestedAt: { type: string, format: date-time }
        acceptedAt:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        hardEndAt:
          type: string
          format: date-time
          nullable: true
        endedAt:
          type: string
          format: date-time
          nullable: true
        endedReason:
          type: string
          nullable: true
        declineReason:
          type: string
          nullable: true
        createdAt: { type: string, format: date-time }
    RequestPrivateSessionInput:
      type: object
      properties:
        requestedMinutes: { type: integer, minimum: 1 }
        note: { type: string, maxLength: 300 }
    DeclinePrivateSessionInput:
      type: object
      properties:
        reason: { type: string, maxLength: 300 }

    # ── Media ─────────────────────────────────────────────────────────────────
    MediaAssetDto:
      type: object
      required: [id, ownerUserId, type, url, status, source, createdAt]
      properties:
        id: { type: string }
        ownerUserId: { type: string }
        creatorId:
          type: string
          nullable: true
        roomId:
          type: string
          nullable: true
        type:
          type: string
          enum: [AVATAR, LOGO, ROOM_COVER, ROOM_THUMBNAIL_CAPTURE, BANNER]
        url: { type: string }
        blurhash:
          type: string
          nullable: true
        dominantColor:
          type: string
          nullable: true
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, REMOVED, HIDDEN]
        source:
          type: string
          enum: [UPLOADED, CREATOR_CAPTURED]
        createdAt: { type: string, format: date-time }

    # ── Reports ───────────────────────────────────────────────────────────────
...
paths:
...
  /creator/rooms/{roomId}/private-sessions:
    get:
      operationId: getCreatorPrivateSessions
      summary: Get private sessions for the creator's room
      tags: [private-sessions]
      parameters:
        - in: path
          name: roomId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: List of private sessions
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/PrivateSessionDto' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /rooms/{roomId}/private-sessions/request:
    post:
      operationId: requestPrivateSession
      summary: Viewer requests a private session in a room
      tags: [private-sessions]
      parameters:
        - in: path
          name: roomId
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequestPrivateSessionInput' }
      responses:
        '201':
          description: Private session requested
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [privateSession, wallet]
                    properties:
                      privateSession: { $ref: '#/components/schemas/PrivateSessionDto' }
                      wallet: { $ref: '#/components/schemas/WalletSummary' }
        '400':
          description: Insufficient tokens or already has active session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /creator/private-sessions/{sessionId}/accept:
    post:
      operationId: acceptPrivateSession
      summary: Creator accepts a private session request
      tags: [private-sessions]
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Private session accepted
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [privateSession]
                    properties:
                      privateSession: { $ref: '#/components/schemas/PrivateSessionDto' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /creator/private-sessions/{sessionId}/decline:
    post:
      operationId: declinePrivateSession
      summary: Creator declines a private session request
      tags: [private-sessions]
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeclinePrivateSessionInput' }
      responses:
        '200':
          description: Private session declined and hold released
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [privateSession]
                    properties:
                      privateSession: { $ref: '#/components/schemas/PrivateSessionDto' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /private-sessions/{sessionId}/start:
    post:
      operationId: startPrivateSession
      summary: Start an accepted private session and get LiveKit credentials
      tags: [private-sessions]
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Private session started
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [privateSession, livekitToken, livekitUrl]
                    properties:
                      privateSession: { $ref: '#/components/schemas/PrivateSessionDto' }
                      livekitToken: { type: string }
                      livekitUrl: { type: string }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /private-sessions/{sessionId}/end:
    post:
      operationId: endPrivateSession
      summary: End an active private session — captures elapsed tokens, releases remainder
      tags: [private-sessions]
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Private session ended
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [privateSession]
                    properties:
                      privateSession: { $ref: '#/components/schemas/PrivateSessionDto' }
                      wallet:
                        type: object
                        allOf:
                          - { $ref: '#/components/schemas/WalletSummary' }
                        nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ── Media ─────────────────────────────────────────────────────────────────────
  /media/upload:
    post:
      operationId: uploadMedia
      summary: Upload avatar, logo, banner, or room cover (multipart)
      tags: [media]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              nullable: true
              type: object
              required: [file, type]
              properties:
                file:
                  type: string
                  format: binary
                type:
                  type: string
                  enum: [AVATAR, LOGO, BANNER, ROOM_COVER]
                creatorId: { type: string }
                roomId: { type: string }
      responses:
        '201':
          description: Media uploaded (status PENDING until reviewed)
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/MediaAssetDto' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: File type or size not accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /media/{mediaId}:
    get:
      operationId: getMedia
      summary: Get media redirect
      tags: [media]
      security: []
      parameters:
        - in: path
          name: mediaId
          required: true
          schema: { type: string }
      responses:
        '302':
          description: Redirect
        '404':
          $ref: '#/components/responses/NotFound'
  /rooms/{roomId}/thumbnail/capture:
    post:
      operationId: captureRoomThumbnail
      summary: Submit a browser-captured frame as the room thumbnail
      tags: [media]
      parameters:
        - in: path
          name: roomId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              nullable: true
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Thumbnail captured and set on room
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [media, room]
                    properties:
                      media: { $ref: '#/components/schemas/MediaAssetDto' }
                      room: { $ref: '#/components/schemas/RoomDetail' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ── Creator earnings ──────────────────────────────────────────────────────────
...